UrlInfo widget returns list of user's uploads - fixed bug in upload.dart button
